+++ mug/__init__.py
@@ -1,3 +1,3 @@
 """Model Upload Guard."""
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"